services/grpc-service/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
